requirements.txt
beautifulsoup4==4.6.0
echarts-python==0.1.3
Flask==0.12.2
jieba==0.38
pymongo==3.4.0
requests==2.18.1

